requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jwcrypto==0.6.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
regex==2020.1.8
requests==2.22.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
